Refactor unique_identifier_cpp
* Change namespace
* Remove Python references
* Update maintainer
* Rename test files
