+++ unique_identifier_cpp/package.xml
@@ -1,30 +1,26 @@
 <?xml version='1.0' encoding='utf-8'?>
 <package format="3">
 
-  <name>unique_id</name>
+  <name>unique_identifier_cpp</name>
   <version>1.0.6</version>
   <description>
     ROS Python and C++ interfaces for universally unique identifiers.
   </description>
-  <maintainer email="jack.oquin@gmail.com">Jack O'Quin</maintainer>
-  <author>Jack O'Quin</author>
+  <maintainer email="jacob@openrobotics.org">Jacob Perron</maintainer>
   <license>BSD</license>
-  <url>http://ros.org/wiki/unique_id</url>
+  <url>https://github.com/ros2/unique_identifier</url>
 
-  <buildtool_depend>ament_cmake_python</buildtool_depend>
+  <author>Jack O'Quin</author>
+
   <buildtool_depend>ament_cmake_ros</buildtool_depend>
 
-  <build_depend>rclpy</build_depend>
   <build_depend>unique_identifier_msgs</build_depend>
 
-  <exec_depend>rclpy</exec_depend>
   <exec_depend>unique_identifier_msgs</exec_depend>
 
   <test_depend>ament_cmake_gtest</test_depend>
-  <test_depend>ament_cmake_nose</test_depend>
 
   <export>
-    <rosdoc config="rosdoc.yaml" />
     <build_type>ament_cmake</build_type>
   </export>
